requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
kedro==0.18.1
openai
setuptools
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
azure-ai-textanalytics==5.2.0b4
Authlib==1.1.0
jsonschema
tenacity